imp(middleware-body): folder renaming: serializers -> parsers
