+++ packages/middleware-body/src/body.middleware.ts
@@ -1,7 +1,7 @@
 import { HttpError, HttpStatus, HttpMiddlewareEffect } from '@marblejs/core';
 import { of, throwError, iif } from 'rxjs';
 import { catchError, map, switchMap, tap, mapTo } from 'rxjs/operators';
-import { defaultParser } from './serializers';
+import { defaultParser } from './parsers';
 import { RequestBodyParser } from './body.model';
 import { matchType, getBody, hasBody } from './body.util';
