+++ packages/middleware-body/src/index.ts
@@ -1,3 +1,3 @@
 export { bodyParser$ } from './body.middleware';
 export { RequestBodyParser } from './body.model';
-export { defaultParser, jsonParser, urlEncodedParser, textParser, rawParser } from './serializers';
+export { defaultParser, jsonParser, urlEncodedParser, textParser, rawParser } from './parsers';
